\name{javadoc}
\Rdversion{1.1}
\alias{javadoc}
\title{Java API of RemoteREngine}
\description{Link to the java API of the RemoteREngine package}
\usage{
javadoc(package = "RemoteREngine", browse = TRUE)
}
\arguments{
	\item{package}{package}
	\item{browse}{logical; if TRUE then the index file of the 
		documentation is opened with \code{\link[utils]{browseURL}} }
}
\value{
Nothing of interest. The function has the side effect of displaying 
the java API of the package. 
}
\author{
Romain Francois <francoisromain@free.fr>
}
\seealso{\link[RemoteREngine:../javadoc/index]{Direct link to java api}}
\examples{\dontrun{
javadoc()
}}
\keyword{ manip }
