\name{RemoteREngine-package}
\Rdversion{1.1}
\alias{RemoteREngine-package}
\alias{RemoteREngine}
\docType{package}
\title{
RMI based implementation of the REngine API 
}
\description{
RMI based implementation of the REngine API
}
\details{
\tabular{ll}{
Package: \tab RemoteREngine\cr
Type: \tab Package\cr
Version: \tab 0.0-4\cr
Date: \tab 2009-09-17\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Romain Francois 

Maintainer: Romain Francois <francoisromain@free.fr>
}
\references{
	REngine API: \url{http://www.rosuda.org/r/nightly/javadoc/index.html?org/rosuda/REngine/REngine.html}
	
	\link[RemoteREngine:../javadoc/index]{RemoteREngine API}
}
\keyword{ package }

